refactor(profile): post-review cleanup for host environment list
Followups from review of the [claude] -> generic environment refactor:

- Drop stale "~ is expanded" claim from Config.environment doc. Behaviour
  was reverted in d977e9e to match sandbox.environment; the comment was
  not.
- Move expand_tilde back to path_input.rs. It was lifted into
  session::environment to share with the original Claude resolver; after
  d977e9e dropped tilde expansion from host_environment_prefix, the only
  caller left is the new-session dialog's path field. Brings the two
  tests along with it.
- Replace the misleading "~/.claude-accounts/work" example in
  test_environment_override_round_trips with an absolute path so the
  test doesn't suggest tilde expansion works at spawn.
- Note in the Host Environment TUI description that all entry forms
  (including bare KEY) resolve to a literal KEY=value arg on the
  spawned process and are visible in `ps`. Recommend sandbox.environment
  for secrets that should stay out of argv.
- Document the new top-level `environment` in
  docs/guides/configuration.md, mirroring the existing
  `sandbox.environment` section.
